Atividades Semtre 1, Nivel Iniciante, Aventureiro, Mestre.

+++ Semestre 01.K3/Aula01.K2.c
@@ -2,13 +2,13 @@
  
 int main() {
     int regularNumber = 2147483647; // Valor máximo de int
-    long int bigNumber = 2147483647;
+     long long int bigNumber = 2147483647;
  
     printf("Número regular (int): %d\n", regularNumber);
     printf("Número grande (long int): %ld\n", bigNumber);
  
     bigNumber = 2147483648; // Valor maior que o máximo de int
-    printf("Número grande atualizado (long int): %ld\n", bigNumber);
+    printf("Número grande atualizado (long int): %lld\n", bigNumber);
  
     return 0;
 }
